+++ game/economies/SimulationResult.java
@@ -3,6 +3,7 @@ import game.economies.FederalEconomy;
 import game.economies.StateEconomy;
 import java.util.*;
 
+//This object contains the federal and state economy results each month
 public class SimulationResult {
     public Map<String, StateEconomy> province;
     public FederalEconomy federation;